+++ app/upgrades/v23/constants.go
@@ -1,6 +1,7 @@
 package v23
 
 import (
+	ibcwasmtypes "github.com/cosmos/ibc-go/modules/light-clients/08-wasm/types"
 	ibcfeetypes "github.com/cosmos/ibc-go/v8/modules/apps/29-fee/types"
 
 	"cosmossdk.io/store/types"
@@ -17,7 +18,9 @@ var Upgrade = upgrades.Upgrade{
 	UpgradeName:          UpgradeName,
 	CreateUpgradeHandler: CreateUpgradeHandler,
 	StoreUpgrades: types.StoreUpgrades{
-		Added:   nil,
+		Added: []string{
+			ibcwasmtypes.StoreKey,
+		},
 		Renamed: nil,
 		Deleted: []string{
 			ibcfeetypes.StoreKey,